[Bug Fixed] STRIDE is still 1 for double-precision.
STRIDE would be added to the data pointer no matter what type the
pointer is. Therefore, data + 1 is the next element of data regardless
of the type of data. (single or double precision)
Pass double-precision tests.

+++ example1.cpp
@@ -2,7 +2,7 @@
 #include <device_matrix.h>
 
 using namespace std;
-typedef device_matrix<float> mat;
+typedef device_matrix<double> mat;
 
 template <typename T>
 void randomInit(device_matrix<T>& m) {
